User can now choose a subset of frames to convert to a video
I added a view function called convert_subset() to ar_gif.py
that accepts a post request specifying a start_frame, end_frame,
basename, and target. It calls the select_subset_of_frames()
function and deletes all frames on the filesystem that
are not part of the subset.

I moved the code that rebinds the convert button to two seperate
functions so that I could easily rebind in different parts of the
code. I now rebind the convert button to call the subset view
after the user has chosen an end_frame.

I split the clear_all() function into two functions.
clear_target() and clear_subset() to allow for more
granular control within the code.